Делит на модули, добавляет (obj, index) => createPost(index))
